ADD: Todo about the kernel swap
From the comment: It might be necessary to swap the kernel here (rotate by 180°),
as the original SignalConv2D function in the python implementation
has a "corr=True" flag, which actually implements correlation instead
of convolution. This would have to be checked with the automated tests
and the reference implementation.

To read more about the difference:
https://towardsdatascience.com/convolution-vs-correlation-af868b6b4fb5